Move types to their own file. Implement parsing of redirect= mechanism
